build.bash update

+++ build.bash
@@ -2,21 +2,36 @@
 # This is a bash file to help you build the project
 #This will create the binary and place it in the /usr/bin folder and set the environment variable so you only need to write Linecounter in your terminal
 
-command -v python3 >/dev/null 2>&1 || { 
-    sudo apt-get update
-    sudo apt-get install python3.6
- }
+function installPy {
+    echo "Do you want to install python 3? (y/n)"
+    read varname
+    if [$varname == "y"]
+    then
+        sudo apt update
+        sudo apt install python3-pip
+    if
+}
 
 command -v python >/dev/null 2>&1 || { 
-    sudo apt update
-    sudo apt install python3-pip
- }
-
+    installPy
+    
+}
+command -v python3 >/dev/null 2>&1 || { 
+    installPy
+}
 
 
 pip install pyinstaller
 
-pyinstaller --onefile ./src/main.py
+PYINSTALLER=$(sudo find / -xdev -name "pyinstaller")
+
+echo $PYINSTALLER
+
+ read -s -n 1 -p "Press any key to continue . . ."
+ echo ""
+
+$PYINSTALLER --onefile ./src/main.py
+
 
 sudo cp ./dist/main /usr/bin/linecounter
